Stabilize local servlets with async processing
This update officially moves the local web server to Thin. There is no
longer an option to use WEBRick. MavensMate.app ajax calls now submit a
request to perform a long running process and receive an async process
id. There is a status servlet that will eventually return the payload
of the response to be handled by the client. This should address the
issues with the local server prematurely closing the connection with
the client on longer running processes.

Refs #25
Resolves #43, #51