Run Active Apex Tests, Payload parameters
New commands to allow for the active or selected files to be
preselected in the Apex Test UI. Only visible when in an MM project and
active when an Apex class with "@isTest" and "testMethod" is in the
selection.

Remove dirs parameter from commands that don't use it.

Update to util.submit_payload() to focus on parameter types instead of
operation to reduce duplication.

Error checking that active file exists on filesystem.

Conflicts:
	util.py