Bug Fix: Linux - Cannot open project #219
This is a fix for the bug #219. The popen command was not passed the parameters in the correct order so the command would fail to execute. This would have been noticed immediately if the application had been tested on Linux. Also, popen takes a list of arguments as the first parameter which is handled better than passing in one single string.